Remove redundant class AuditedObject.
